requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
pyyaml==6.0.1


